+++ lib/src/animated.dart
@@ -1,20 +1,13 @@
 import 'package:flutter/widgets.dart';
 
 import 'ticked.dart';
-import 'initialized.dart';
+import 'disposed.dart';
 
 /// An helper widget that instanciate an [AnimationController] and dispose it.
 class Animated extends StatelessWidget {
-  final Duration duration;
-  final Duration reverseDuration;
-  final String debugLabel;
-  final double lowerBound;
-  final double upperBound;
-  final double value;
-  final AnimationBehavior animationBehavior;
-  final InitializedWidgetBuilder<AnimationController> builder;
   const Animated({
-    Key key,
+    Key? key,
+    required this.builder,
     this.value,
     this.duration,
     this.reverseDuration,
@@ -22,9 +15,16 @@ class Animated extends StatelessWidget {
     this.lowerBound,
     this.upperBound,
     this.animationBehavior,
-    @required this.builder,
-  })  : assert(builder != null),
-        super(key: key);
+  }) : super(key: key);
+
+  final Duration? duration;
+  final Duration? reverseDuration;
+  final String? debugLabel;
+  final double? lowerBound;
+  final double? upperBound;
+  final double? value;
+  final AnimationBehavior? animationBehavior;
+  final DisposedWidgetBuilder<AnimationController> builder;
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +40,7 @@ class Animated extends StatelessWidget {
           value: value,
           vsync: vsync,
         );
-        return Disposed(
+        return DisposedValue(
           value: result,
           dispose: () => result.dispose(),
         );