+++ lib/src/paged.dart
@@ -1,33 +1,32 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import 'initialized.dart';
+import 'disposed.dart';
 
 /// An helper widget that instanciate an [PageController] and dispose it.
 class Paged extends StatelessWidget {
-  final InitializedWidgetBuilder<PageController> builder;
-  final int initialPage;
-  final double viewportFraction;
-  final bool keepPage;
+  final DisposedWidgetBuilder<PageController> builder;
+  final int? initialPage;
+  final double? viewportFraction;
+  final bool? keepPage;
   const Paged({
-    Key key,
+    Key? key,
     this.initialPage,
     this.keepPage,
     this.viewportFraction,
-    @required this.builder,
-  })  : assert(builder != null),
-        super(key: key);
+    required this.builder,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Initialized<PageController>(
+    return Disposed<PageController>(
       initialize: (context) {
         final result = PageController(
           initialPage: initialPage ?? 0,
           keepPage: keepPage ?? true,
           viewportFraction: viewportFraction ?? 1.0,
         );
-        return Disposed(
+        return DisposedValue(
           value: result,
           dispose: () => result.dispose(),
         );