+++ lib/src/text_edited.dart
@@ -1,24 +1,23 @@
 import 'package:flutter/widgets.dart';
 
-import 'initialized.dart';
+import 'disposed.dart';
 
 /// An helper widget that instanciate an [TextEditingController] and dispose it.
 class TextEdited extends StatelessWidget {
   final String text;
-  final InitializedWidgetBuilder<TextEditingController> builder;
+  final DisposedWidgetBuilder<TextEditingController> builder;
   const TextEdited({
-    Key key,
+    Key? key,
     this.text = '',
-    @required this.builder,
-  })  : assert(builder != null),
-        super(key: key);
+    required this.builder,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Initialized<TextEditingController>(
+    return Disposed<TextEditingController>(
       initialize: (context) {
         final result = TextEditingController(text: text);
-        return Disposed(
+        return DisposedValue(
           value: result,
           dispose: () => result.dispose(),
         );