+++ lib/src/ticked.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 
-import 'initialized.dart';
+import 'disposed.dart';
 
-typedef TickedInitializer<T> = Disposed<T> Function(
+typedef TickedInitializer<T> = DisposedValue<T> Function(
   BuildContext context,
   TickerProvider vsync,
 );
@@ -11,32 +11,30 @@ typedef TickedInitializer<T> = Disposed<T> Function(
 /// [initState] and [dispose] it.
 class Ticked<T> extends StatefulWidget {
   final TickedInitializer<T> initialize;
-  final InitializedWidgetBuilder<T> builder;
+  final DisposedWidgetBuilder<T> builder;
   const Ticked({
-    Key key,
-    this.initialize,
-    @required this.builder,
-  })  : assert(builder != null),
-        super(key: key);
+    Key? key,
+    required this.initialize,
+    required this.builder,
+  }) : super(key: key);
 
   @override
   _TickedState<T> createState() => _TickedState<T>();
 }
 
 class _TickedState<T> extends State<Ticked<T>> with TickerProviderStateMixin {
-  Disposed<T> value;
+  late DisposedValue<T> value;
 
   @override
   void initState() {
-    if (widget.initialize != null) {
-      value = widget.initialize(context, this);
-    }
+    value = widget.initialize(context, this);
+
     super.initState();
   }
 
   @override
   void dispose() {
-    value?.dispose?.call();
+    value.dispose?.call();
     super.dispose();
   }