+++ lib/stateful.dart
@@ -1,7 +1,7 @@
 library stateful;
 
 export 'src/animated.dart';
-export 'src/initialized.dart';
+export 'src/disposed.dart';
 export 'src/paged.dart';
 export 'src/registered.dart';
 export 'src/registered_extensions.dart';
@@ -9,3 +9,4 @@ export 'src/scrolled.dart';
 export 'src/tabbed.dart';
 export 'src/text_edited.dart';
 export 'src/ticked.dart';
+export 'src/initializer.dart';