+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.2.0] - Initial version
+
+* Migrated to null safety.
+* Renamed `Initialized` to `Disposed`
+* Added the `Initializer`
+
 ## [0.1.0] - Initial version
 
 * First stateful helpers.